Split LLMTTS panel into dedicated LLM and TTS panels
