--- web/src/panels/perception/LLMTTSPanel.css
@@ -1,24 +1,0 @@
-/* Tab row — panel-specific */
-.hri-tab-row {
-  display: flex;
-  gap: var(--space-1);
-}
-
-.hri-tab-btn {
-  flex: 1;
-  padding: var(--space-1) 0;
-  background: var(--surface-raised);
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius);
-  color: var(--text-muted);
-  font-size: var(--font-size-sm);
-  font-weight: 600;
-  cursor: pointer;
-  transition: background 0.12s, border-color 0.12s, color 0.12s;
-}
-.hri-tab-btn:hover { color: var(--text-primary); border-color: var(--border-bright); }
-.hri-tab-btn.active {
-  background: var(--color-info-bg);
-  border-color: var(--color-info);
-  color: var(--color-info);
-}
